[telemetry] Rename export facet to exporting
Rename the export folder and namespace to exporting to avoid confusion
with the C++ export keyword which required using 'exp' as the namespace
name. Now the folder and namespace names are aligned.

Changes:
- Rename folders: export/ -> exporting/ (include and src)
- Update namespace: ores::telemetry::exp -> ores::telemetry::exporting
- Update all header guards, include paths, and namespace declarations
- Update consumers in ores.shell (parser.cpp, options.hpp, host.cpp)
- Update modeling documentation
- Add net to standard facets in component-creator skill
