+++ projects/ores.shell/include/ores.shell/config/options.hpp
@@ -25,7 +25,7 @@
 #include "ores.shell/config/login_options.hpp"
 #include "ores.comms/net/client_options.hpp"
 #include "ores.telemetry/log/logging_options.hpp"
-#include "ores.telemetry/export/telemetry_options.hpp"
+#include "ores.telemetry/exporting/telemetry_options.hpp"
 
 namespace ores::shell::config {
 
@@ -51,7 +51,7 @@ struct options final {
     /**
      * @brief Configuration options for telemetry export, if any.
      */
-    std::optional<telemetry::exp::telemetry_options> telemetry;
+    std::optional<telemetry::exporting::telemetry_options> telemetry;
 };
 
 std::ostream& operator<<(std::ostream& s, const options& v);
