+++ projects/ores.shell/src/config/parser.cpp
@@ -26,7 +26,7 @@
 #include "ores.shell/config/parser_exception.hpp"
 #include "ores.utility/version/version.hpp"
 #include "ores.telemetry/log/logging_configuration.hpp"
-#include "ores.telemetry/export/telemetry_configuration.hpp"
+#include "ores.telemetry/exporting/telemetry_configuration.hpp"
 #include "ores.utility/program_options/environment_mapper_factory.hpp"
 
 namespace {
@@ -59,7 +59,7 @@ using ores::shell::config::parser_exception;
  */
 options_description make_options_description() {
     using ores::telemetry::log::logging_configuration;
-    using ores::telemetry::exp::telemetry_configuration;
+    using ores::telemetry::exporting::telemetry_configuration;
 
     options_description god("General");
     god.add_options()
@@ -187,7 +187,7 @@ read_login_configuration(const variables_map& vm) {
 std::optional<options>
 parse_arguments(const std::vector<std::string>& arguments, std::ostream& info) {
     using ores::telemetry::log::logging_configuration;
-    using ores::telemetry::exp::telemetry_configuration;
+    using ores::telemetry::exporting::telemetry_configuration;
     using ores::utility::program_options::environment_mapper_factory;
 
     const auto od(make_options_description());
